commited

+++ src/app/features/story/story-table/story-table.component.ts
@@ -9,65 +9,70 @@ import { SharedService } from '../../../core/services/shared.service';
 import { Sprint } from '../../../core/models/sprint.model';
 import { TaskService } from '../../../core/services/task.service';
 import { Task } from '../../../core/models/task.model';
+import { DndDropEvent } from 'ngx-drag-drop';
 import { TaskFormComponent } from '../task-form/task-form.component';
-import { MatMenuModule } from '@angular/material/menu';
-import { MatIconModule } from '@angular/material/icon';
-import { MatButtonModule } from '@angular/material/button';
-import { Route, Router } from '@angular/router';
+import { Router } from '@angular/router';
 import { EmployeeService } from '../../../core/services/employee.service';
 import { Employee } from '../../../core/models/employee.model';
+import { LoaderService } from '../../../core/services/loader.service';
 
 @Component({
   selector: 'app-story-table',
-  standalone: false,
   templateUrl: './story-table.component.html',
   styleUrls: ['./story-table.component.scss'],
+  standalone:false,
   encapsulation: ViewEncapsulation.None,
 })
 export class StoryTableComponent implements OnInit {
   storyList: Story[] = [];
   sprintList: Sprint[] = [];
   taskList: Task[] = [];
-  employeeList:Employee[]=[];
+  employeeList: Employee[] = [];
   hoveredTaskId: number | null = null;
   selectedSprint!: Sprint;
+  draggedTask: any;
+  draggedStory: any;
+  draggedColumn: string='';
 
   constructor(
     private shared: SharedService,
-    private employee:EmployeeService,
+    private employee: EmployeeService,
     private story: StoryService,
     private alert: AlertService,
     private dialog: MatDialog,
     private task: TaskService,
     private router: Router,
-    private sprint: SprintService
-  ) {}
+    private sprint: SprintService,
+    private loader: LoaderService
+  ) { }
 
   ngOnInit(): void {
     this.getAllSprint();
     this.changeInDropdown();
     this.getEmployeeList();
   }
-  getEmployeeList()
-  {
+
+  getEmployeeList() {
+    this.loader.show();
     this.employee.getEmployees().subscribe({
-      next:(res)=>{
-        this.employeeList=res;
+      next: (res) => {
+        this.employeeList = res;
+        this.loader.hide();
       },
-      error:(err)=>{
-        this.alert.sidePopUp('Yo got some error while fetching Employees','error');
-        console.log('error',err)
+      error: (err) => {
+        this.alert.sidePopUp(err.message, 'error');
+        this.loader.hide();
       }
-    })
+    });
   }
-  getEmployeeName(id:number|undefined):string|null{
-    const emp=this.employeeList.find((e)=>String(e.id)==String(id));
-    return emp?emp.name:null;
+
+  getEmployeeName(id: number | undefined): string | null {
+    const emp = this.employeeList.find((e) => String(e.id) == String(id));
+    return emp ? emp.name : null;
   }
 
   changeInDropdown() {
     this.shared.selectedProjectId$.subscribe((projectId) => {
-      console.log('Project ID changed:', projectId);
       if (projectId) {
         this.getAllSprint();
       }
@@ -79,37 +84,67 @@ export class StoryTableComponent implements OnInit {
     return tasks.filter((task) => task.status === status);
   }
 
- getAllSprint() {
-  const id = this.shared.getSelectedProjectId() || '';
-  const prevSelectedSprintId = this.selectedSprint?.id;
+  getAllSprint() {
+    this.loader.show();
+    const id = this.shared.getSelectedProjectId() || '';
+    const prevSelectedSprintId = this.selectedSprint?.id;
 
-  this.sprint.getSprint(id).subscribe({
-    next: (res: Sprint[]) => {
-      this.sprintList = res;
-
-      if (res.length > 0) {
-        const matchedSprint = res.find(s => s.id === prevSelectedSprintId);
-
-        if (matchedSprint) {
-          this.selectedSprint = matchedSprint;
-        } else {
-          this.selectedSprint = res[0]; 
+    this.sprint.getSprint(id).subscribe({
+      next: (res: Sprint[]) => {
+        this.sprintList = res;
+        if (res.length > 0) {
+          const matchedSprint = res.find(s => s.id === prevSelectedSprintId);
+          this.selectedSprint = matchedSprint ? matchedSprint : res[0];
+          this.getAllStory(this.selectedSprint.id.toString());
         }
-
-        this.getAllStory(this.selectedSprint.id.toString());
-      }
-    },
-    error: (err) => {
-      this.alert.sidePopUp('Error fetching sprint data', 'error');
-      console.log('this is error in sprint ', err);
-    },
-  });
-}
+        this.loader.hide();
+      },
+      error: (err) => {
+        this.alert.sidePopUp('Error fetching sprint data', 'error');
+        this.loader.hide();
+      },
+    });
+  }
 
   onSprintChange(event: any) {
-    console.log('Sprint changed:', this.selectedSprint);
     this.getAllStory(event.id);
   }
+
+  onDragStart(event: DragEvent, task: any, story: any, column: string) {
+    this.draggedTask = task;
+    this.draggedStory = story;
+    this.draggedColumn = column;
+    event.dataTransfer?.setData('text/plain', JSON.stringify({taskId: task.id}));
+    console.log("Dragged task:", this.draggedTask, "from story:", this.draggedStory, "in column:", this.draggedColumn);
+  }
+
+  onDragOver(event: DragEvent) {
+    console.log("This is sragged task",event);
+    event.preventDefault();
+  }
+  onDrop(event: DragEvent, story: any, newColumn: string) {
+  event.preventDefault();
+
+  if (this.draggedTask && this.draggedStory === story) {
+    if (this.draggedTask.status !== newColumn) {
+      const payload = { ...this.draggedTask, status: newColumn };
+      this.task.editStatusTask(this.draggedTask.id.toString(), payload).subscribe({
+        next: () => {
+          this.getAllStory(this.selectedSprint.id.toString());
+        },
+        error: (err: any) => {
+          this.alert.sidePopUp(err.message, 'error');
+        }
+      });
+    } else {
+      this.alert.sidePopUp('Task already in ' + newColumn + ' column', 'warning');
+    }
+  }
+
+  this.draggedTask = null;
+  this.draggedStory = null;
+  this.draggedColumn = '';
+}
 
   getAllStory(sprintId: string) {
     this.story.getStory(sprintId).subscribe({
@@ -134,14 +169,12 @@ export class StoryTableComponent implements OnInit {
               if (loaded === stories.length) {
                 this.storyList = stories;
               }
-              console.error('Error loading tasks for story', story.id, err);
             },
           });
         });
       },
       error: (err) => {
-        this.alert.sidePopUp('You got some error', 'error');
-        console.log('error', err);
+        this.alert.sidePopUp(err.message, 'error');
         this.storyList = [];
       },
     });
@@ -165,13 +198,18 @@ export class StoryTableComponent implements OnInit {
   }
 
   async deleteDialog(event: Story) {
-     const confirmed = await this.alert.confirmDelete();
-    this.story.deleteStory(event.id.toString()).subscribe({
-      next: () => {
-        this.alert.sidePopUp('This record is deleted', 'success');
-        this.getAllSprint();
-      },
-    });
+    const confirmed = await this.alert.confirmDelete();
+    if (confirmed) {
+      this.story.deleteStory(event.id.toString()).subscribe({
+        next: () => {
+          this.alert.sidePopUp('This record is deleted', 'success');
+          this.getAllSprint();
+        },
+        error: (err) => {
+          this.alert.sidePopUp(err.message, 'error');
+        }
+      });
+    }
   }
 
   addTask(event: Story) {
@@ -192,25 +230,32 @@ export class StoryTableComponent implements OnInit {
 
   async onDeleteTask(t: Task) {
     const confirmed = await this.alert.confirmDelete();
-    if(confirmed){
-    this.task.deleteTask(t.id.toString()).subscribe({
-      next: () => {
-        this.getAllSprint();
-      },
-      error: (err) => {
-        console.log('the error while deleting', err);
-      },
-    });}
+    if (confirmed) {
+      this.task.deleteTask(t.id.toString()).subscribe({
+        next: () => {
+          this.getAllSprint();
+        },
+        error: (err) => {
+          this.alert.sidePopUp(err.message, 'error');
+        },
+      });
+    }
   }
-  loggingEffort(task: Task): void {
-  this.router.navigate(['scrum-board/logging'], {
-    queryParams: {
-      taskId: task.id,
-      sprintId: this.selectedSprint?.id 
-    },
-  });
-}
 
+  loggingEffort(task: Task): void {
+    const id = this.shared.getUserId();
+    if (id == task.assign_to?.toString()) {
+      this.router.navigate(['scrum-board/logging'], {
+        queryParams: {
+          taskId: task.id,
+          sprintId: this.selectedSprint?.id
+        },
+      });
+    }
+    else {
+      this.alert.sidePopUp('You are not authorized to log efforts for this task', 'error');
+    }
+  }
 
   onEditTask(task: Task) {
     this.dialog
